Extracted Bottom Bar to separate file
